More attendance adjustments.  Realized I left a hole in where the attendance was still seeping in to the localized table through officer notes.  This has been corrected.  Also the localized table gets reapplied when the timer is stopped now.  Just incase the notes hadn't finished updating when you stop the timer, it will prevent partial guild attendance ticks at the end.
This happened because a guild roster update fires after the timer is off which allows the officer note and attendance to be updated from officer notes.  If this happens while the attendance is being written, it will cause the table it is reading from to be overwritten with what is in the officer notes.

Also added LT_NameIsRaider() return t/f and LT_RankIsRaider return t/f.  Consider removing all hard coded ranks with this in your if statements so we can start using ranks as options rather than hard coded.